pancake-order-system/status_service/requirements.txt
redis
python-dotenv 
redis
python-dotenv
flask
flask-cors